Simplify reminder id flow to int callback
Agent-Logs-Url: https://github.com/Metatronius/work_in_progress/sessions/574b1f5d-0f08-4874-b3ee-f4e2bdbf2935

+++ app/src/main/java/com/example/work_in_progress/database/TaskViewModel.kt
@@ -23,7 +23,7 @@ class TaskViewModel(private val taskRepository: TaskRepository) : ViewModel() {
      *
      * @param newTask The parameter object containing the task fields to persist.
      */
-    fun addTask(newTask: TaskParams, onInserted: ((Long) -> Unit)? = null) {
+    fun addTask(newTask: TaskParams, onInserted: ((Int) -> Unit)? = null) {
         require(newTask.title.isNotBlank() && newTask.title.length in 0..30) { "Title must not be blank or exceed 30 characters." }
         if (newTask.due !== null)
             DataUtil.validateDate(newTask.due)
@@ -40,7 +40,7 @@ class TaskViewModel(private val taskRepository: TaskRepository) : ViewModel() {
                     target = 1
                 )
             )
-            onInserted?.invoke(insertedId)
+            onInserted?.invoke(Math.toIntExact(insertedId))
         }
     }
